Move useClientId hook to hooks folder
