+++ packages/lesswrong/components/common/AttributionInViewTracker.tsx
@@ -5,7 +5,7 @@ import { useCurrentUser } from './withUser';
 import { RecombeeViewPortionProps, recombeeApi } from '../../lib/recombee/client';
 import { recombeeEnabledSetting } from '@/lib/instanceSettings';
 import { isRecombeeRecommendablePost } from '@/lib/collections/posts/helpers';
-import { useClientId } from '@/lib/abTestImpl';
+import { useClientId } from '../hooks/useClientId';
 
 interface AttributionEventProps {
   post: PostsListBase;
