+++ packages/lesswrong/lib/abTestImpl.ts
@@ -2,8 +2,6 @@
 
 import React from 'react';
 import rng from './seedrandom';
-import { CLIENT_ID_COOKIE } from './cookies/cookies';
-import { useCookiesWithConsent } from '../components/hooks/useCookiesWithConsent';
 import { randomId } from './random';
 import keyBy from 'lodash/keyBy';
 import { allABTests } from './abTests';
@@ -134,19 +132,6 @@ export function weightedRandomPick<T extends string>(options: Record<T,number>, 
       return key;
   }
   throw new Error("Out of range value in weightedRandomPick");
-}
-
-// Returns the user's clientID. This is stored in a cookie separately from
-// accounts; a user may have multiple clientIDs (eg if they have multiple
-// devices) and a clientID may correspond to multiple users (if they log out and
-// log in with a different account).
-//
-// A logged-out user's client ID determines which A/B test groups they are in.
-// A logged-in user has their A/B test groups determined by the client ID they
-// had when they created their account.
-export function useClientId(): string | undefined {
-  const [cookies] = useCookiesWithConsent([CLIENT_ID_COOKIE]);
-  return cookies[CLIENT_ID_COOKIE];
 }
 
 export function classesForAbTestGroups(groups: CompleteTestGroupAllocation) {
